Address review feedback: fix Int64 overflow, remove unused import, improve idempotency
- Use explicit Int64 literal for 60_000_000_000 division to avoid 32-bit overflow
- Remove unused `import Foundation` from HeaderHelpers+SyntheticDurationStubs.swift
- Update registerDurationFactoryExtensionProperty to also refresh property type
  and getter accessor (signature + external link name) when re-registering

+++ Sources/Runtime/RuntimeDuration.swift
@@ -85,7 +85,7 @@ public func kk_duration_inWholeMinutes(_ durationRaw: Int) -> Int {
     guard let box = runtimeDurationBox(from: durationRaw) else {
         fatalError("KSwiftK panic [\(runtimePanicDiagnosticCode)]: kk_duration_inWholeMinutes received invalid Duration handle")
     }
-    return Int(box.nanoseconds / 60_000_000_000)
+    return Int(box.nanoseconds / Int64(60_000_000_000))
 }
 
 @_cdecl("kk_duration_inWholeNanoseconds")
